chore: refactor tests into smaller files
Before the tests were in monolithic files: test_momento.py and
test_momento_async.py. This refactors the tests into smaller files
based on behavior: config, init, control API, scalar API.
