+++ tests/momento/simple_cache_client/test_scalar.py
@@ -1,235 +1,12 @@
-import os
 import time
 from datetime import timedelta
 
-import pytest
-
-import momento.errors as errors
 from momento import SimpleCacheClient
-from momento.auth.credential_provider import CredentialProvider, EnvMomentoTokenProvider
+from momento.auth.credential_provider import EnvMomentoTokenProvider
 from momento.config.configuration import Configuration
-from momento.errors import InvalidArgumentException, MomentoErrorCode
-from momento.responses import (
-    CacheDeleteResponse,
-    CacheGetResponse,
-    CacheSetResponse,
-    CreateCacheResponse,
-    DeleteCacheResponse,
-    ListCachesResponse,
-)
-from tests.utils import str_to_bytes, unique_test_cache_name, uuid_bytes, uuid_str
-
-
-def test_create_cache_get_set_values_and_delete_cache(client: SimpleCacheClient, cache_name: str) -> None:
-    random_cache_name = unique_test_cache_name()
-    key = uuid_str()
-    value = uuid_str()
-
-    client.create_cache(random_cache_name)
-
-    set_resp = client.set(random_cache_name, key, value)
-    assert isinstance(set_resp, CacheSetResponse.Success)
-
-    get_resp = client.get(random_cache_name, key)
-    assert isinstance(get_resp, CacheGetResponse.Hit)
-    assert get_resp.value_string == value
-
-    get_for_key_in_some_other_cache = client.get(cache_name, key)
-    assert isinstance(get_for_key_in_some_other_cache, CacheGetResponse.Miss)
-
-    client.delete_cache(random_cache_name)
-
-
-# Init
-def test_init_throws_exception_when_client_uses_negative_default_ttl(
-    configuration: Configuration, credential_provider: CredentialProvider
-) -> None:
-    with pytest.raises(InvalidArgumentException, match="TTL timedelta must be a non-negative integer"):
-        SimpleCacheClient(configuration, credential_provider, timedelta(seconds=-1))
-
-
-def test_init_throws_exception_for_non_jwt_token(configuration: Configuration, default_ttl_seconds: timedelta) -> None:
-    with pytest.raises(InvalidArgumentException, match="Invalid Auth token."):
-        os.environ["BAD_AUTH_TOKEN"] = "notanauthtoken"
-        credential_provider = EnvMomentoTokenProvider("BAD_AUTH_TOKEN")
-        SimpleCacheClient(configuration, credential_provider, default_ttl_seconds)
-
-
-def test_init_throws_exception_when_client_uses_integer_request_timeout_ms(
-    configuration: Configuration, credential_provider: CredentialProvider, default_ttl_seconds: int
-) -> None:
-    with pytest.raises(
-        InvalidArgumentException, match="Request timeout must be a timedelta with a value greater " "than zero."
-    ):
-        configuration.with_client_timeout(-1)
-
-
-def test_init_throws_exception_when_client_uses_negative_request_timeout_ms(
-    configuration: Configuration, credential_provider: CredentialProvider, default_ttl_seconds: timedelta
-) -> None:
-    with pytest.raises(
-        InvalidArgumentException, match="Request timeout must be a timedelta with a value greater than zero."
-    ):
-        configuration = configuration.with_client_timeout(timedelta(seconds=-1))
-        SimpleCacheClient(configuration, credential_provider, default_ttl_seconds)
-
-
-def test_init_throws_exception_when_client_uses_zero_request_timeout_ms(
-    configuration: Configuration, credential_provider: CredentialProvider, default_ttl_seconds: timedelta
-) -> None:
-    with pytest.raises(
-        InvalidArgumentException, match="Request timeout must be a timedelta with a value greater than zero."
-    ):
-        configuration = configuration.with_client_timeout(timedelta(seconds=0))
-        SimpleCacheClient(configuration, credential_provider, default_ttl_seconds)
-
-
-# Create cache
-def test_create_cache__already_exists_when_creating_existing_cache(client: SimpleCacheClient, cache_name: str) -> None:
-    response = client.create_cache(cache_name)
-    assert isinstance(response, CreateCacheResponse.CacheAlreadyExists)
-
-
-def test_create_cache_throws_exception_for_empty_cache_name(
-    client: SimpleCacheClient,
-) -> None:
-    response = client.create_cache("")
-    assert isinstance(response, CreateCacheResponse.Error)
-    assert response.error_code == MomentoErrorCode.INVALID_ARGUMENT_ERROR
-
-
-def test_create_cache_throws_validation_exception_for_null_cache_name(
-    client: SimpleCacheClient,
-) -> None:
-    response = client.create_cache(None)
-    assert isinstance(response, CreateCacheResponse.Error)
-    assert response.error_code == MomentoErrorCode.INVALID_ARGUMENT_ERROR
-    assert response.inner_exception.message == "Cache name must be a non-empty string"
-
-
-def test_create_cache_with_bad_cache_name_throws_exception(
-    client: SimpleCacheClient,
-) -> None:
-    response = client.create_cache(1)
-    assert isinstance(response, CreateCacheResponse.Error)
-    assert response.error_code == MomentoErrorCode.INVALID_ARGUMENT_ERROR
-    assert response.inner_exception.message == "Cache name must be a non-empty string"
-
-
-def test_create_cache_throws_authentication_exception_for_bad_token(
-    bad_token_credential_provider: EnvMomentoTokenProvider, configuration: Configuration, default_ttl_seconds: timedelta
-) -> None:
-    with SimpleCacheClient(configuration, bad_token_credential_provider, default_ttl_seconds) as client:
-        response = client.create_cache(unique_test_cache_name())
-        assert isinstance(response, CreateCacheResponse.Error)
-        assert response.error_code == errors.MomentoErrorCode.AUTHENTICATION_ERROR
-
-
-# Delete cache
-def test_delete_cache_succeeds(client: SimpleCacheClient, cache_name: str) -> None:
-    cache_name = uuid_str()
-
-    response = client.create_cache(cache_name)
-    assert isinstance(response, CreateCacheResponse.Success)
-
-    response = client.delete_cache(cache_name)
-    assert isinstance(response, DeleteCacheResponse.Success)
-
-    response = client.delete_cache(cache_name)
-    assert isinstance(response, DeleteCacheResponse.Error)
-    assert response.error_code == MomentoErrorCode.NOT_FOUND_ERROR
-
-
-def test_delete_cache_throws_not_found_when_deleting_unknown_cache(
-    client: SimpleCacheClient,
-) -> None:
-    cache_name = uuid_str()
-    response = client.delete_cache(cache_name)
-    assert isinstance(response, DeleteCacheResponse.Error)
-    assert response.error_code == MomentoErrorCode.NOT_FOUND_ERROR
-
-
-def test_delete_cache_throws_invalid_input_for_null_cache_name(
-    client: SimpleCacheClient,
-) -> None:
-    response = client.delete_cache(None)
-    assert isinstance(response, DeleteCacheResponse.Error)
-    assert response.error_code == MomentoErrorCode.INVALID_ARGUMENT_ERROR
-
-
-def test_delete_cache_throws_exception_for_empty_cache_name(
-    client: SimpleCacheClient,
-) -> None:
-    response = client.delete_cache("")
-    assert isinstance(response, DeleteCacheResponse.Error)
-    assert response.error_code == MomentoErrorCode.INVALID_ARGUMENT_ERROR
-
-
-def test_delete_with_bad_cache_name_throws_exception(client: SimpleCacheClient, cache_name: str) -> None:
-    response = client.delete_cache(1)
-    assert isinstance(response, DeleteCacheResponse.Error)
-    assert response.error_code == MomentoErrorCode.INVALID_ARGUMENT_ERROR
-    assert response.inner_exception.message == "Cache name must be a non-empty string"
-
-
-def test_delete_cache_throws_authentication_exception_for_bad_token(
-    bad_token_credential_provider: EnvMomentoTokenProvider, configuration: Configuration, default_ttl_seconds: timedelta
-) -> None:
-    with SimpleCacheClient(configuration, bad_token_credential_provider, default_ttl_seconds) as client:
-        response = client.delete_cache(uuid_str())
-        assert isinstance(response, DeleteCacheResponse.Error)
-        assert response.error_code == MomentoErrorCode.AUTHENTICATION_ERROR
-
-
-# List caches
-def test_list_caches_succeeds(client: SimpleCacheClient, cache_name: str) -> None:
-    cache_name = uuid_str()
-
-    initial_response = client.list_caches()
-    assert isinstance(initial_response, ListCachesResponse.Success)
-
-    cache_names = [cache.name for cache in initial_response.caches]
-    assert cache_name not in cache_names
-
-    try:
-        response = client.create_cache(cache_name)
-        assert isinstance(response, CreateCacheResponse.Success)
-
-        list_cache_resp = client.list_caches()
-        assert isinstance(list_cache_resp, ListCachesResponse.Success)
-
-        cache_names = [cache.name for cache in list_cache_resp.caches]
-        assert cache_name in cache_names
-        assert list_cache_resp.next_token is None
-    finally:
-        delete_response = client.delete_cache(cache_name)
-        assert isinstance(delete_response, DeleteCacheResponse.Success)
-
-
-def test_list_caches_throws_authentication_exception_for_bad_token(
-    bad_token_credential_provider: EnvMomentoTokenProvider, configuration: Configuration, default_ttl_seconds: timedelta
-) -> None:
-    with SimpleCacheClient(configuration, bad_token_credential_provider, default_ttl_seconds) as client:
-        response = client.list_caches()
-        assert isinstance(response, ListCachesResponse.Error)
-        assert response.error_code == MomentoErrorCode.AUTHENTICATION_ERROR
-
-
-def test_list_caches_with_next_token_works(client: SimpleCacheClient, cache_name: str) -> None:
-    """skip until pagination is actually implemented, see
-    https://github.com/momentohq/control-plane-service/issues/83"""
-    pass
-
-
-# Signing keys
-def test_create_list_revoke_signing_keys(client: SimpleCacheClient) -> None:
-    create_resp = client.create_signing_key(timedelta(minutes=30))
-    list_resp = client.list_signing_keys()
-    assert create_resp.key_id() in [signing_key.key_id() for signing_key in list_resp.signing_keys()]
-
-    client.revoke_signing_key(create_resp.key_id())
-    list_resp = client.list_signing_keys()
-    assert create_resp.key_id() not in [signing_key.key_id() for signing_key in list_resp.signing_keys()]
+from momento.errors import MomentoErrorCode
+from momento.responses import CacheDeleteResponse, CacheGetResponse, CacheSetResponse
+from tests.utils import str_to_bytes, uuid_bytes, uuid_str
 
 
 # Setting and Getting
@@ -499,13 +276,3 @@ def test_delete(client: SimpleCacheClient, cache_name: str) -> None:
     # Verify deleted
     get_response = client.get(cache_name, key)
     assert isinstance(get_response, CacheGetResponse.Miss)
-
-
-def test_configuration_client_timeout_copy_constructor(configuration: Configuration) -> None:
-    def snag_deadline(config: Configuration) -> timedelta:
-        return config.get_transport_strategy().get_grpc_configuration().get_deadline()  # type: ignore
-
-    original_deadline: timedelta = snag_deadline(configuration)
-    assert original_deadline.total_seconds() == 15
-    configuration = configuration.with_client_timeout(timedelta(seconds=600))
-    assert snag_deadline(configuration).total_seconds() == 600
